+++ packages/ionic/src/commands/cordova/prepare.ts
@@ -22,7 +22,7 @@ You may wish to use ${chalk.green('ionic cordova prepare')} if you run your proj
   inputs: [
     {
       name: 'platform',
-      description: `The platform you would like to prepare (e.g. ${chalk.green('ios')}, ${chalk.green('android')})`,
+      description: `The platform you would like to prepare (${['android', 'ios'].map(v => chalk.green(v)).join(', ')})`,
       required: false,
     },
   ]