+++ packages/ionic/src/commands/cordova/resources.ts
@@ -45,7 +45,7 @@ This command uses Ionic servers, so we require you to be logged into your free I
   inputs: [
     {
       name: 'platform',
-      description: `The platform for which you would like to generate resources (e.g. ${chalk.green('ios')}, ${chalk.green('android')})`,
+      description: `The platform for which you would like to generate resources (${['android', 'ios'].map(v => chalk.green(v)).join(', ')})`,
       required: false,
     }
   ],