+++ packages/ionic/src/commands/cordova/run.ts
@@ -19,7 +19,7 @@ Just like with ${chalk.green('ionic cordova build')}, you can pass additional op
   inputs: [
     {
       name: 'platform',
-      description: `The platform to run: ${chalk.green('ios')}, ${chalk.green('android')}`,
+      description: `The platform to run (${['android', 'ios'].map(v => chalk.green(v)).join(', ')})`,
     }
   ],
   options: CORDOVA_RUN_COMMAND_OPTIONS,