+++ CHANGELOG.md
@@ -17,6 +17,7 @@ the CHANGELOG files within CLI packages. For specific commit-level changes, see
 * Added app-scripts arguments (such as `--prod`) to `ionic upload`.
 * Added SSL configuration options for CLI HTTP requests. See
   [README.md#ssl-configuration](https://github.com/ionic-team/ionic-cli/blob/master/README.md#ssl-configuration).
+* Added Cordova arguments to `ionic cordova compile`.
 * Fixed issue with `ionic serve` requiring an external network.
 * Fixed `--consolelogs` not working for Ionic 1.
 * Fixed issue with `ionic cordova platform add` not using custom icon and